--- Pipeline/playground/spectrum.py
@@ -1,26 +1,0 @@
-import librosa
-import librosa.display
-import matplotlib.pyplot as plt
-import numpy as np
-
-#audio_file = "audio/kepler STEMS BASS.mp3"
-audio_file = "../audio/kepler STEMS DRUMS.mp3"
-
-y, sr = librosa.load(audio_file, offset=25, duration=5, sr=None)
-fig, ax = plt.subplots(nrows=2, ncols=1, sharex=True)
-D = librosa.amplitude_to_db(np.abs(librosa.stft(y)), ref=np.max)
-img = librosa.display.specshow(D, y_axis='linear', x_axis='time',
-                               sr=sr, ax=ax[0])
-ax[0].set(title='Linear-frequency power spectrogram')
-ax[0].label_outer()
-
-hop_length = 1024
-D = librosa.amplitude_to_db(np.abs(librosa.stft(y, hop_length=hop_length)),
-                            ref=np.max)
-librosa.display.specshow(D, y_axis='log', sr=sr, hop_length=hop_length,
-                         x_axis='time', ax=ax[1])
-ax[1].set(title='Log-frequency power spectrogram')
-ax[1].label_outer()
-fig.colorbar(img, ax=ax, format="%+2.f dB")
-
-plt.show()
