+++ playground/bandwidth.py
@@ -1,0 +1,28 @@
+import matplotlib.pyplot as plt
+import librosa.display
+import matplotlib.pyplot as plt
+import numpy as np
+
+audio_file = "../audio/kepler STEMS BASS.mp3"
+
+y, sr = librosa.load(audio_file, offset=19, duration=15)
+spec_bw = librosa.feature.spectral_bandwidth(y=y, sr=sr)
+S, phase = librosa.magphase(librosa.stft(y=y))
+
+fig, ax = plt.subplots(nrows=2, sharex=True)
+times = librosa.times_like(spec_bw)
+centroid = librosa.feature.spectral_centroid(S=S)
+ax[0].semilogy(times, spec_bw[0], label='Spectral bandwidth')
+ax[0].set(ylabel='Hz', xticks=[], xlim=[times.min(), times.max()])
+ax[0].legend()
+ax[0].label_outer()
+librosa.display.specshow(librosa.amplitude_to_db(S, ref=np.max),
+                         y_axis='log', x_axis='time', ax=ax[1])
+ax[1].set(title='log Power spectrogram')
+ax[1].fill_between(times, np.maximum(0, centroid[0] - spec_bw[0]),
+                np.minimum(centroid[0] + spec_bw[0], sr/2),
+                alpha=0.5, label='Centroid +- bandwidth')
+ax[1].plot(times, centroid[0], label='Spectral centroid', color='w')
+ax[1].legend(loc='lower right')
+
+plt.show()
