+++ playground/mfcc.py
@@ -1,0 +1,42 @@
+import librosa
+import librosa.display
+import matplotlib.pyplot as plt
+import numpy as np
+
+audio_file = "kepler.mp3"
+
+y, sr = librosa.load(audio_file)
+
+S = librosa.feature.melspectrogram(y=y, sr=sr, n_mels=128,
+                                   fmax=8000)
+
+mfccs = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=40)
+
+fig, ax = plt.subplots(nrows=2, sharex=True)
+img = librosa.display.specshow(librosa.power_to_db(S, ref=np.max),
+                               x_axis='time', y_axis='mel', fmax=8000,
+                               ax=ax[0])
+
+fig.colorbar(img, ax=[ax[0]])
+ax[0].set(title='Mel spectrogram')
+ax[0].label_outer()
+img = librosa.display.specshow(mfccs, x_axis='time', ax=ax[1])
+fig.colorbar(img, ax=[ax[1]])
+ax[1].set(title='MFCC')
+
+
+m_slaney = librosa.feature.mfcc(y=y, sr=sr, dct_type=2)
+m_htk = librosa.feature.mfcc(y=y, sr=sr, dct_type=3)
+fig, ax = plt.subplots(nrows=2, sharex=True, sharey=True)
+img1 = librosa.display.specshow(m_slaney, x_axis='time', ax=ax[0])
+ax[0].set(title='RASTAMAT / Auditory toolbox (dct_type=2)')
+fig.colorbar(img, ax=[ax[0]])
+img2 = librosa.display.specshow(m_htk, x_axis='time', ax=ax[1])
+ax[1].set(title='HTK-style (dct_type=3)')
+fig.colorbar(img2, ax=[ax[1]])
+
+#plt.imshow(img.reshape((28, 28)))
+plt.show()
+
+print("Hello World")
+
