+++ playground/onset.py
@@ -1,0 +1,26 @@
+import matplotlib.pyplot as plt
+import librosa.display
+import scipy.signal
+import matplotlib.pyplot as plt
+import numpy as np
+
+audio_file = "../audio/kepler STEMS DRUMS.mp3"
+
+y, sr = librosa.load(audio_file, offset=26, duration=15)
+librosa.onset.onset_detect(y=y, sr=sr, units='time')
+o_env = librosa.onset.onset_strength(y=y, sr=sr)
+times = librosa.times_like(o_env, sr=sr)
+onset_frames = librosa.onset.onset_detect(onset_envelope=o_env, sr=sr)
+
+D = np.abs(librosa.stft(y))
+fig, ax = plt.subplots(nrows=2, sharex=True)
+librosa.display.specshow(librosa.amplitude_to_db(D, ref=np.max),
+                         x_axis='time', y_axis='log', ax=ax[0])
+ax[0].set(title='Power spectrogram')
+ax[0].label_outer()
+ax[1].plot(times, o_env, label='Onset strength')
+ax[1].vlines(times[onset_frames], 0, o_env.max(), color='r', alpha=0.9,
+           linestyle='--', label='Onsets')
+ax[1].legend()
+
+plt.show()
