+++ sonosthesia-audio-pipeline/__init__.py
@@ -1,0 +1,3 @@
+
+def hello_world():
+    print("Hello world")
