+++ sonosthesia-audio-pipeline/__main__.py
@@ -1,0 +1,5 @@
+def hello_world():
+    print("Hello world")
+
+if __name__ == '__main__':
+    hello_world()
